fix(gameplay): restore MegaMek hex angle geometry and remove generic ECM to-hit modifier
