+++ src/utils/gameplay/toHit/c3.ts
@@ -16,7 +16,7 @@ import {
   isBetterBracket,
 } from '../c3Network';
 import { IWeaponRangeProfile } from '../range';
-import { calculateToHit, type IEcmContext } from './calculate';
+import { calculateToHit } from './calculate';
 
 export interface IC3ToHitInput {
   readonly attackerEntityId: string;
@@ -99,7 +99,7 @@ export function calculateToHitWithC3(
   range: number,
   c3Input: IC3ToHitInput,
   minRange: number = 0,
-  weaponIdOrEcmContext?: string | IEcmContext,
+  weaponId?: string,
   semiGuidedTagContext?: ISemiGuidedTagToHitContext,
 ): IToHitCalculation & { readonly c3Result: IC3TargetingResult } {
   const c3Result = getC3TargetingBenefit(
@@ -121,7 +121,7 @@ export function calculateToHitWithC3(
     effectiveBracket,
     range,
     minRange,
-    weaponIdOrEcmContext,
+    weaponId,
     semiGuidedTagContext,
   );
