+++ src/utils/gameplay/toHit/index.ts
@@ -45,11 +45,7 @@ export {
   calculateIndirectFireModifier,
   calculateCalledShotModifier,
 } from './damageModifiers';
-export {
-  calculateToHit,
-  calculateToHitFromContext,
-  type IEcmContext,
-} from './calculate';
+export { calculateToHit, calculateToHitFromContext } from './calculate';
 export {
   calculateSemiGuidedTagIndirectFireModifier,
   calculateSemiGuidedTagTargetMovementModifier,
@@ -62,16 +58,6 @@ export {
 } from './stateHydration';
 export { calculateToHitWithC3 } from './c3';
 export type { IC3ToHitInput } from './c3';
-
-// Per `add-ecm-tohit-modifier` (closes playtest gap #1).
-export {
-  calculateEcmModifier,
-  ECM_TO_HIT_MODIFIERS,
-  ECM_MODIFIER_VALUE,
-  type EcmModifierReason,
-  type IEcmCoverageState,
-  type WeaponGuidanceType,
-} from './ecmModifier';
 
 export {
   buildToHitForecast,
